Refactor the IsExpired Method
The IsExpired method should be refactored and relocated from the
AstarteFailedMessageEntry class to the AstarteFailedMessageStorage
class. This refactoring entails revising the logic used to calculate
the expiry time of messages. The absolute expiry time is determined by
 the moment when the device stores fallback messages in the database.

The IsExpired method will then compare the timestamp of the message
in the database with the time when the device attempts to
resend messages. If the absolute expiry time is greater
than the time of message sending, the messages will be removed
from the database.

Signed-off-by: Osman Hadzic <osman.hadzic@secomind.com>